// src/services/generated/radarr.ts
export interface components {
    schemas: {
// ...
        DiskSpaceResource: {
            /** Format: int32 */
            id?: number;
            path?: string | null;
            label?: string | null;
            /** Format: int64 */
            freeSpace?: number;
            /** Format: int64 */
            totalSpace?: number;
        };
// ...
        HealthResource: {
            /** Format: int32 */
            id?: number;
            source?: string | null;
            type?: components["schemas"]["HealthCheckResult"];
            message?: string | null;
            wikiUrl?: string | null;
        };
// ...
        SystemResource: {
            appName?: string | null;
            instanceName?: string | null;
            version?: string | null;
            /** Format: date-time */
            buildTime?: string;
            isDebug?: boolean;
            isProduction?: boolean;
            isAdmin?: boolean;
            isUserInteractive?: boolean;
            startupPath?: string | null;
            appData?: string | null;
            osName?: string | null;
            osVersion?: string | null;
            isNetCore?: boolean;
            isLinux?: boolean;
            isOsx?: boolean;
            isWindows?: boolean;
            isDocker?: boolean;
            mode?: components["schemas"]["RuntimeMode"];
            branch?: string | null;
            databaseType?: components["schemas"]["DatabaseType"];
            databaseVersion?: string | null;
            authentication?: components["schemas"]["AuthenticationType"];
            /** Format: int32 */
            migrationVersion?: number;
            urlBase?: string | null;
            runtimeVersion?: string | null;
            runtimeName?: string | null;
            /** Format: date-time */
            startTime?: string;
            packageVersion?: string | null;
            packageAuthor?: string | null;
            packageUpdateMechanism?: components["schemas"]["UpdateMechanism"];
            packageUpdateMechanismMessage?: string | null;
        };
// ...
        TaskResource: {
            /** Format: int32 */
            id?: number;
            name?: string | null;
            taskName?: string | null;
            /** Format: int32 */
            interval?: number;
            /** Format: date-time */
            lastExecution?: string;
            /** Format: date-time */
            lastStartTime?: string;
            /** Format: date-time */
            nextExecution?: string;
            /** Format: date-span */
            readonly lastDuration?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}